+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
 - Prevent crash when filename is not provided (fixes #81)
 - Update the call to the Record resource to use named attributes. (#97)
 - Show detailed error when data is not posted with multipart content-type.
+- Fix crash when submitted data is not valid JSON (fixes #104)
 
 **Internal changes**